Chore: make pages'
