+++ src/App.tsx
@@ -1,27 +1,10 @@
 import "./App.css";
-import { icons } from "./components/icons";
-import NavBar from "@components/NavBar";
-import { Button } from "@/components/ui/button";
+// import { icons } from "./components/icons";
+import HomePage from "@/pages/home/homepage";
+
 
 function App() {
-  const navItems = [
-    {
-      label: "Dashboard",
-      link: "/dashboard",
-    },
-    {
-      label: "API-services",
-      link: "/apiservices",
-    },
-    {
-      label: "Pricing",
-      link: "/pricing",
-    },
-    {
-      label: "Profile",
-      link: "/profile",
-    },
-  ];
+  
   return (
     <>
       {/* <h1 className="text-3xl font-bold underline">Hello world!</h1>
@@ -30,8 +13,7 @@ function App() {
       <icons.upload />
       <icons.loop /> */}
 
-      <NavBar logoText="MyanSen" navItems={navItems} />
-      <Button variant="outline">Button</Button>
+     <HomePage/>
     </>
   );
 }
